requirements.txt
openai-whisper>=1.3.0
tqdm>=4.66.1
torch>=2.2.1